+++ src/CG.Blazor/ServiceCollectionExtensions.cs
@@ -3,7 +3,7 @@ using CG.Validations;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace CG.Blazor
+namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
     /// This class contains extension methods related to the <see cref="IServiceCollection"/>