+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * I added the TokenProvider type.
 * I added the ITokenService and TokenServiceBase types
 * I added the IStateService and StateService types.
+* I changed the namespace for the ServiceCollectionExtension class to Microsoft.Extensions.DependencyInjection
 
 ## 2021.1